Add files via upload
Adds an OP "Fuel Replicator" called the 691-FR. With the press of a button, it will (hopefully) exchange 1 Electric charge with 200 of all other fuel types

(including electric charge, talk about hacks). I'm going to be adding more sizes in the future, and hopefully adding this to some fuel tanks for mobile fuel replication.